Extract `LeapActivity` and `LeapState` to separate modules
