+++ node/src/components/sync_leaper.rs
@@ -1,17 +1,13 @@
 //! The Sync Leaper
 mod error;
 mod event;
+mod leap_activity;
+mod leap_state;
 mod metrics;
 #[cfg(test)]
 mod tests;
 
-use std::{
-    cmp::Ordering,
-    collections::HashMap,
-    fmt::{Display, Formatter},
-    sync::Arc,
-    time::Instant,
-};
+use std::{sync::Arc, time::Instant};
 
 use datasize::DataSize;
 use prometheus::Registry;
@@ -28,206 +24,21 @@ use crate::{
 };
 pub(crate) use error::LeapActivityError;
 pub(crate) use event::Event;
+pub(crate) use leap_state::LeapState;
 
 use metrics::Metrics;
+
+use self::leap_activity::LeapActivity;
 
 const COMPONENT_NAME: &str = "sync_leaper";
 
 #[derive(Clone, Debug, DataSize)]
-enum PeerState {
+pub(crate) enum PeerState {
     RequestSent,
     Rejected,
     CouldntFetch,
     Fetched(Box<SyncLeap>),
 }
-
-#[derive(Debug, DataSize)]
-pub(crate) enum LeapState {
-    Idle,
-    Awaiting {
-        sync_leap_identifier: SyncLeapIdentifier,
-        in_flight: usize,
-    },
-    Received {
-        best_available: Box<SyncLeap>,
-        from_peers: Vec<NodeId>,
-        in_flight: usize,
-    },
-    Failed {
-        sync_leap_identifier: SyncLeapIdentifier,
-        error: LeapActivityError,
-        from_peers: Vec<NodeId>,
-        in_flight: usize,
-    },
-}
-
-impl Display for LeapState {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        match self {
-            LeapState::Idle => {
-                write!(f, "Idle")
-            }
-            LeapState::Awaiting {
-                sync_leap_identifier,
-                in_flight,
-            } => {
-                write!(
-                    f,
-                    "Awaiting {} responses for {}",
-                    in_flight,
-                    sync_leap_identifier.block_hash(),
-                )
-            }
-            LeapState::Received {
-                best_available,
-                from_peers,
-                in_flight,
-            } => {
-                write!(
-                    f,
-                    "Received {} from {} peers, awaiting {} responses",
-                    best_available.highest_block_hash(),
-                    from_peers.len(),
-                    in_flight
-                )
-            }
-            LeapState::Failed {
-                sync_leap_identifier,
-                error,
-                ..
-            } => {
-                write!(
-                    f,
-                    "Failed leap for {} {}",
-                    sync_leap_identifier.block_hash(),
-                    error
-                )
-            }
-        }
-    }
-}
-
-impl LeapState {
-    fn in_flight(&self) -> usize {
-        match self {
-            LeapState::Idle => 0,
-            LeapState::Awaiting { in_flight, .. }
-            | LeapState::Received { in_flight, .. }
-            | LeapState::Failed { in_flight, .. } => *in_flight,
-        }
-    }
-
-    fn active(&self) -> bool {
-        self.in_flight() > 0
-    }
-}
-
-#[derive(Debug, DataSize)]
-struct LeapActivity {
-    sync_leap_identifier: SyncLeapIdentifier,
-    peers: HashMap<NodeId, PeerState>,
-    leap_start: Instant,
-}
-
-impl LeapActivity {
-    fn status(&self) -> LeapState {
-        let sync_leap_identifier = self.sync_leap_identifier;
-        let in_flight = self
-            .peers
-            .values()
-            .filter(|state| matches!(state, PeerState::RequestSent))
-            .count();
-        let responsed = self.peers.len() - in_flight;
-
-        // TODO[RC]: Should this be just self.peers.is_empty()?
-
-        if in_flight == 0 && responsed == 0 {
-            return LeapState::Failed {
-                sync_leap_identifier,
-                in_flight,
-                error: LeapActivityError::NoPeers(sync_leap_identifier),
-                from_peers: vec![],
-            };
-        }
-        if in_flight > 0 && responsed == 0 {
-            return LeapState::Awaiting {
-                sync_leap_identifier,
-                in_flight,
-            };
-        }
-        match self.best_response() {
-            Ok((best_available, from_peers)) => LeapState::Received {
-                in_flight,
-                best_available: Box::new(best_available),
-                from_peers,
-            },
-            // `Unobtainable` means we couldn't download it from any peer so far - don't treat it
-            // as a failure if there are still requests in flight
-            Err(LeapActivityError::Unobtainable(_, _)) if in_flight > 0 => LeapState::Awaiting {
-                sync_leap_identifier,
-                in_flight,
-            },
-            Err(error) => LeapState::Failed {
-                sync_leap_identifier,
-                from_peers: vec![],
-                in_flight,
-                error,
-            },
-        }
-    }
-
-    fn best_response(&self) -> Result<(SyncLeap, Vec<NodeId>), LeapActivityError> {
-        let reject_count = self
-            .peers
-            .values()
-            .filter(|peer_state| matches!(peer_state, PeerState::Rejected))
-            .count();
-
-        let mut peers = vec![];
-        let mut maybe_ret: Option<&Box<SyncLeap>> = None;
-        for (peer, peer_state) in &self.peers {
-            match peer_state {
-                PeerState::Fetched(sync_leap) => match &maybe_ret {
-                    None => {
-                        maybe_ret = Some(sync_leap);
-                        peers.push(*peer);
-                    }
-                    Some(current_ret) => {
-                        match current_ret
-                            .highest_block_height()
-                            .cmp(&sync_leap.highest_block_height())
-                        {
-                            Ordering::Less => {
-                                maybe_ret = Some(sync_leap);
-                                peers = vec![*peer];
-                            }
-                            Ordering::Equal => {
-                                peers.push(*peer);
-                            }
-                            Ordering::Greater => {}
-                        }
-                    }
-                },
-                PeerState::RequestSent | PeerState::Rejected | PeerState::CouldntFetch => {}
-            }
-        }
-
-        match maybe_ret {
-            Some(sync_leap) => Ok((*sync_leap.clone(), peers)),
-            None => {
-                if reject_count > 0 {
-                    Err(LeapActivityError::TooOld(self.sync_leap_identifier, peers))
-                } else {
-                    Err(LeapActivityError::Unobtainable(
-                        self.sync_leap_identifier,
-                        peers,
-                    ))
-                }
-            }
-        }
-    }
-}
-
 #[derive(Debug, DataSize)]
 pub(crate) struct SyncLeaper {
     leap_activity: Option<LeapActivity>,
@@ -259,7 +70,7 @@ impl SyncLeaper {
                         LeapState::Received { .. } | LeapState::Failed { .. } => {
                             self.metrics
                                 .sync_leap_duration
-                                .observe(activity.leap_start.elapsed().as_secs_f64());
+                                .observe(activity.leap_start().elapsed().as_secs_f64());
                         }
                         LeapState::Idle | LeapState::Awaiting { .. } => {
                             // should be unreachable
@@ -289,9 +100,9 @@ impl SyncLeaper {
             return effects;
         }
         if let Some(leap_activity) = self.leap_activity.as_mut() {
-            if leap_activity.sync_leap_identifier != sync_leap_identifier {
+            if leap_activity.sync_leap_identifier() != &sync_leap_identifier {
                 error!(
-                    current_sync_leap_identifier = %leap_activity.sync_leap_identifier,
+                    current_sync_leap_identifier = %leap_activity.sync_leap_identifier(),
                     requested_sync_leap_identifier = %sync_leap_identifier,
                     "tried to start fetching a sync leap for a different sync_leap_identifier"
                 );
@@ -299,7 +110,7 @@ impl SyncLeaper {
             }
 
             for peer in peers_to_ask {
-                if false == leap_activity.peers.contains_key(&peer) {
+                if false == leap_activity.peers().contains_key(&peer) {
                     effects.extend(
                         effect_builder
                             .fetch::<SyncLeap>(sync_leap_identifier, peer, self.chainspec.clone())
@@ -308,7 +119,9 @@ impl SyncLeaper {
                                 fetch_result,
                             }),
                     );
-                    leap_activity.peers.insert(peer, PeerState::RequestSent);
+                    leap_activity
+                        .peers_mut()
+                        .insert(peer, PeerState::RequestSent);
                 }
             }
             return effects;
@@ -328,11 +141,11 @@ impl SyncLeaper {
                 (peer, PeerState::RequestSent)
             })
             .collect();
-        self.leap_activity = Some(LeapActivity {
+        self.leap_activity = Some(LeapActivity::new(
             sync_leap_identifier,
             peers,
-            leap_start: Instant::now(),
-        });
+            Instant::now(),
+        ));
         effects
     }
 
@@ -352,9 +165,9 @@ impl SyncLeaper {
             }
         };
 
-        if leap_activity.sync_leap_identifier != sync_leap_identifier {
+        if leap_activity.sync_leap_identifier() != &sync_leap_identifier {
             warn!(
-                requested_hash=%leap_activity.sync_leap_identifier,
+                requested_hash=%leap_activity.sync_leap_identifier(),
                 response_hash=%sync_leap_identifier,
                 "block hash in the response doesn't match the one requested"
             );
@@ -366,7 +179,7 @@ impl SyncLeaper {
                 error!(%sync_leap_identifier, "fetched a sync leap from storage - should never happen");
             }
             Ok(FetchedData::FromPeer { item, peer, .. }) => {
-                let peer_state = match leap_activity.peers.get_mut(&peer) {
+                let peer_state = match leap_activity.peers_mut().get_mut(&peer) {
                     Some(state) => state,
                     None => {
                         warn!(
@@ -381,7 +194,7 @@ impl SyncLeaper {
                 self.metrics.sync_leap_fetched_from_peer.inc();
             }
             Err(fetcher::Error::Rejected { peer, .. }) => {
-                let peer_state = match leap_activity.peers.get_mut(&peer) {
+                let peer_state = match leap_activity.peers_mut().get_mut(&peer) {
                     Some(state) => state,
                     None => {
                         warn!(
@@ -399,7 +212,7 @@ impl SyncLeaper {
             Err(error) => {
                 let peer = error.peer();
                 info!(?error, %peer, %sync_leap_identifier, "failed to fetch a sync leap from peer");
-                let peer_state = match leap_activity.peers.get_mut(peer) {
+                let peer_state = match leap_activity.peers_mut().get_mut(peer) {
                     Some(state) => state,
                     None => {
                         warn!(
